Add new STRG+SHIFT+UP/DOWN shortcut for moving sections up/down
This patch adds the shortcuts STRG+SHIFT+UP and STRG+SHIFT+DOWN to move
a section (as defined as having the same heading) up or down, swapping
it with the previous/next section. It currently uses a linear search in
some places where a binary search similiar to the one in
outlinewidget.cpp might be more performant.